pyproject.toml
[project]
name = "job-email-generator"
version = "1.0.0"
description = "AI-powered job application email generator backend"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "fastapi==0.109.0",
    "uvicorn[standard]==0.27.0",
    "python-dotenv==1.0.0",
    "pydantic==2.5.3",
    "sqlalchemy==2.0.25",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "openai>=1.55.0",
    "pgvector==0.2.5",
    "python-multipart==0.0.6",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
]

[project.optional-dependencies]
dev = [
    "requests>=2.31.0",  # For test_api.py
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.uv]
dev-dependencies = [
    "requests>=2.31.0",
]
